+++ lib/src/generators/deities/shrines/deity_few_shrines_generator.dart
@@ -1,0 +1,53 @@
+import '../../base/list_item_generator.dart';
+
+/// A class that generates the amount of shrines a deity has
+class DeityFewShrinesGenerator extends ListItemGenerator<String> {
+  static const _shrines = [
+    "are hard to find",
+    "can only be found in remote places",
+    "don't exist",
+    "hardly exist",
+    "can be accessed only by worshipers",
+    'are hidden underground',
+    'are situated on mountaintops',
+    'are inside living trees',
+    'are built inside caves',
+    'are floating on clouds',
+    'are present only in ruins',
+    'are guarded by magical barriers',
+    'are accessible only during certain times of the year',
+    'are found on islands far from the mainland',
+    'are constructed only in underground tunnels',
+    'are located inside ancient ruins',
+    'are guarded by fierce beasts',
+    'are hidden in plain sight',
+    'are guarded by priests',
+    'are found only in desert oases',
+    'are built in underwater caverns',
+    'are present only in abandoned fortresses',
+    'are built in secret underground tunnels',
+    'are accessible only by a secret password or code',
+    'are constructed in the branches of giant trees',
+    'are accessible only by climbing treacherous cliffs',
+    'are built on the ruins of other shrines',
+    'are present only in deep caverns',
+    'are constructed only on the top of very high mountains',
+    "are guarded by the deity's followers",
+    'are accessible only by sea',
+    'are located at the bottom of the ocean',
+    'are built on top of active volcanoes',
+    'are found only in the midst of vast deserts',
+    'are present only in the homes of the rich and powerful',
+    'are hidden behind waterfalls',
+    'are built inside the trunks of ancient trees',
+    'are found on the backs of giant turtles',
+    'are made of stone and blend in with the landscape',
+    'are accessible only by crossing treacherous swamps',
+    "are present only in the underground lairs of the deity's followers",
+    'are hidden inside the ruins of ancient temples',
+    'are guarded by powerful magical creatures',
+    'are present only in the most remote and desolate places'
+  ];
+
+  DeityFewShrinesGenerator() : super(_shrines);
+}